renamed namespaces part of #110
more explicit hiding and pulling out of stuff
