+++ mypy.ini
@@ -6,7 +6,7 @@
 
 # explicitly blacklist files, this means we can easily add in stricter type checks
 # by removing them from the blacklist
-[mypy-chemcoord.cartesian_coordinates.asymmetric_unit_cartesian_class,chemcoord.cartesian_coordinates._cartesian_class_core,chemcoord.cartesian_coordinates._cartesian_class_get_zmat,chemcoord.cartesian_coordinates._cartesian_class_io,chemcoord.cartesian_coordinates.cartesian_class_main,chemcoord.cartesian_coordinates._cartesian_class_pandas_wrapper,chemcoord.cartesian_coordinates._cartesian_class_symmetry,chemcoord.cartesian_coordinates._cart_transformation,chemcoord.cartesian_coordinates._indexers,chemcoord.cartesian_coordinates.point_group,chemcoord.cartesian_coordinates.xyz_functions,chemcoord.configuration,chemcoord.constants,chemcoord.exceptions,chemcoord._generic_classes.generic_core,chemcoord._generic_classes.generic_IO,chemcoord.internal_coordinates._indexers,chemcoord.internal_coordinates._zmat_class_core,chemcoord.internal_coordinates._zmat_class_io,chemcoord.internal_coordinates.zmat_class_main,chemcoord.internal_coordinates._zmat_class_pandas_wrapper,chemcoord.internal_coordinates.zmat_functions,chemcoord.internal_coordinates._zmat_transformation,chemcoord.utilities._decorators,chemcoord.utilities._print_versions,chemcoord.utilities._temporary_deprecation_workarounds,chemcoord]
+[mypy-chemcoord._cartesian_coordinates.asymmetric_unit_cartesian_class,chemcoord._cartesian_coordinates._cartesian_class_core,chemcoord._cartesian_coordinates._cartesian_class_get_zmat,chemcoord._cartesian_coordinates._cartesian_class_io,chemcoord._cartesian_coordinates.cartesian_class_main,chemcoord._cartesian_coordinates._cartesian_class_pandas_wrapper,chemcoord._cartesian_coordinates._cartesian_class_symmetry,chemcoord._cartesian_coordinates._cart_transformation,chemcoord._cartesian_coordinates._indexers,chemcoord._cartesian_coordinates.point_group,chemcoord._cartesian_coordinates.xyz_functions,chemcoord.configuration,chemcoord.constants,chemcoord.exceptions,chemcoord._generic_classes.generic_core,chemcoord._generic_classes.generic_IO,chemcoord._internal_coordinates._indexers,chemcoord._internal_coordinates._zmat_class_core,chemcoord._internal_coordinates._zmat_class_io,chemcoord._internal_coordinates.zmat_class_main,chemcoord._internal_coordinates._zmat_class_pandas_wrapper,chemcoord._internal_coordinates.zmat_functions,chemcoord._internal_coordinates._zmat_transformation,chemcoord._utilities._decorators,chemcoord._utilities._print_versions,chemcoord._utilities._temporary_deprecation_workarounds,chemcoord]
     disallow_untyped_defs = False
     check_untyped_defs = False
