+++ src/chemcoord/_cartesian_coordinates/_indexers.py
@@ -2,7 +2,7 @@
 
 import warnings
 
-from chemcoord.utilities._temporary_deprecation_workarounds import is_iterable
+from chemcoord._utilities._temporary_deprecation_workarounds import is_iterable
 
 
 class _generic_Indexer(object):
