+++ src/chemcoord/_cartesian_coordinates/asymmetric_unit_cartesian_class.py
@@ -2,7 +2,7 @@
 import numpy as np
 import pandas as pd
 
-from chemcoord.cartesian_coordinates.cartesian_class_main import Cartesian
+from chemcoord._cartesian_coordinates.cartesian_class_main import Cartesian
 
 
 class AsymmetricUnitCartesian(Cartesian):
