+++ src/chemcoord/_cartesian_coordinates/cartesian_class_main.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
-from chemcoord.cartesian_coordinates._cartesian_class_get_zmat import CartesianGetZmat
-from chemcoord.cartesian_coordinates._cartesian_class_io import CartesianIO
-from chemcoord.cartesian_coordinates._cartesian_class_symmetry import CartesianSymmetry
+from chemcoord._cartesian_coordinates._cartesian_class_get_zmat import CartesianGetZmat
+from chemcoord._cartesian_coordinates._cartesian_class_io import CartesianIO
+from chemcoord._cartesian_coordinates._cartesian_class_symmetry import CartesianSymmetry
 
 
 class Cartesian(CartesianIO, CartesianGetZmat, CartesianSymmetry):
