+++ src/chemcoord/_internal_coordinates/_indexers.py
@@ -2,8 +2,8 @@
 
 import warnings
 
+from chemcoord._utilities._temporary_deprecation_workarounds import is_iterable
 from chemcoord.exceptions import InvalidReference
-from chemcoord.utilities._temporary_deprecation_workarounds import is_iterable
 
 
 class _generic_Indexer(object):
