+++ src/chemcoord/_internal_coordinates/_zmat_class_io.py
@@ -5,8 +5,8 @@ import pandas as pd
 
 import chemcoord.constants as constants
 from chemcoord._generic_classes.generic_IO import GenericIO
+from chemcoord._internal_coordinates._zmat_class_core import ZmatCore
 from chemcoord.exceptions import InvalidReference, UndefinedCoordinateSystem
-from chemcoord.internal_coordinates._zmat_class_core import ZmatCore
 
 
 class ZmatIO(ZmatCore, GenericIO):
