+++ src/chemcoord/_internal_coordinates/_zmat_transformation.py
@@ -14,12 +14,12 @@ from numpy import cos, cross, sin
 from numpy.linalg import inv
 
 import chemcoord.constants as constants
-from chemcoord.cartesian_coordinates._cart_transformation import (
+from chemcoord._cartesian_coordinates._cart_transformation import (
     get_B,
     get_grad_B,
     get_ref_pos,
 )
-from chemcoord.cartesian_coordinates.xyz_functions import _jit_isclose
+from chemcoord._cartesian_coordinates.xyz_functions import _jit_isclose
 from chemcoord.exceptions import ERR_CODE_OK, ERR_CODE_InvalidReference
