+++ src/chemcoord/_internal_coordinates/zmat_class_main.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-from chemcoord.internal_coordinates._zmat_class_io import ZmatIO
+from chemcoord._internal_coordinates._zmat_class_io import ZmatIO
 
 # from chemcoord.internal_coordinates._zmat_class_gradient import ZmatGradient
